refactor: recruit_count.ts を common/ に移動
