+++ src/app/feat-recruit/common/recruit_count.ts
@@ -1,5 +1,5 @@
-import { RecruitCountService } from '../../../../db/recruit_count_service';
-import { exists } from '../../../common/others';
+import { RecruitCountService } from '../../../db/recruit_count_service';
+import { exists } from '../../common/others';
 
 export async function increaseRecruitCount(userIdList: string[]) {
     userIdList.forEach(async (userId) => {
